+++ docs/changelog.md
@@ -12,7 +12,7 @@ Sprint 10 / April 12, 2020
 
 - added total daily time to Dashboard.
 - updated readme.
-- added tests for Timer.
+- added tests for Timer and Dashboard.
 - made Form's Input a separate component.
 - removed time.ts utility file.
 - rename Time component's constructStringFromSeconds method to getStringFromTimeUnits.